+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
+- shared: use `five8` to encode/decode ([#70](https://github.com/lamports-dev/richat/pull/70))
+
 ### Breaking
 
 ## 2025-02-11